Added jumping to the Strider.
Right now it just aimlessly wanders and occasionally picks a random spot to jump to, with a random chance of doing a quick jump instead of a normal one.

Of course, this will be used for multiple things, like dodging, backing off, and just generally moving (i.e scaling buildings).