Added basic grenade attack function.
Begun adding the grenade code, so far it's just the preparation animation and firing state, and basic handling for firing at the position the target was last seen. Also improved the grenade projectiles' homing mode.
Next up should be trying to add handling for firing over raised floor walls, and the decision logic for firing the grenade launcher, oh and also giving it a sound.